Annotator processor kapt option 'fritz2.visibility' to specify lens object visibility and default to explicit 'public'
