+++ lenses-annotation-processor/src/main/resources/META-INF/services/javax.annotation.processing.Processor
